fix(data-layer): rework AddonDB initialisation and add tests
The initial implementation clashed with Mikro-orm when mocking during tests. Refactoring this way acted as aworkaround for the issue

+++ src/data-layer/AddonDB.ts
@@ -1,47 +1,36 @@
-import {
-  Connection,
-  EntityManager,
-  EntityName,
-  IDatabaseDriver,
-  MikroORM,
-} from "@mikro-orm/core";
-import { SqlEntityManager, SqliteDriver } from "@mikro-orm/sqlite";
-import { Addon } from "./entities/Addon.entity";
-import { File } from "./entities/File.entity";
-import { join as joinPath } from "path";
+import { SqliteDriver } from "@mikro-orm/sqlite";
 import { IncomingAddon } from "../incoming-entities/incomingAddon";
 import { IncomingFile } from "../incoming-entities/incomingFile";
-import { addonMapper } from "./mappers/addonMapper"
-import { fileMapper } from "./mappers/fileMapper"
-
+import { initOrm } from "./initOrm";
+import { addonMapper } from "./mappers/addonMapper";
+import { fileMapper } from "./mappers/fileMapper";
 
 export class AddonDB {
-  private orm: MikroORM<SqliteDriver>;
-  private entityManager: SqlEntityManager<SqliteDriver> &
-    EntityManager<IDatabaseDriver<Connection>>;
-  public entities;
-  private ormConfig;
+  public orm;
+  public entityManager;
 
-  constructor() {
-    this.ormConfig = {
-      dbName: "addon-db",
-      type: "sqlite",
-      entities: [joinPath("dist", "data-layer", "entities")],
-      entitiesTs: [joinPath("src", "data-layer", "entities")],
-    };
-  }
+  constructor() {}
 
   public async init() {
-    this.orm = await MikroORM.init<SqliteDriver>(this.ormConfig);
+    this.orm = await initOrm();
+    await this.orm.connect()
     this.entityManager = this.orm.em.fork();
     await this.orm.close();
   }
 
-  public async getById(table: EntityName<Addon | File>, id: string) {
+  public async getById(table: any, id: string) {
     await this.orm.connect();
-    const result = await this.entityManager.findOne(table, {
-      entityId: Number(id),
-    });
+
+    let result;
+    try {
+      result = await this.entityManager.findOne(table, {
+        entityId: Number(id),
+      });
+    } catch (error) {
+      console.log(error);
+      return undefined
+    }
+
     await this.orm.close();
     return result;
   }
@@ -53,45 +42,44 @@ export class AddonDB {
     return result;
   }
 
-  public async removeById(table: EntityName<Addon | File>, id: string){
-    const itemToRemove = await this.getById(table, id)
-    await this.orm.connect();
-    this.entityManager.remove(itemToRemove);
-    await this.entityManager.flush()
-    await this.orm.close();
-    return itemToRemove
+  public async removeById(table: any, id: string) {
+    const itemToRemove = await this.getById(table, id);
+    if(itemToRemove){
+      await this.orm.connect();
+      this.entityManager.remove(itemToRemove);
+      await this.entityManager.flush();
+      await this.orm.close();
+      return itemToRemove;
+    } else {
+      return undefined
+    }
   }
 
-  public async writeEntity(type: EntityName<Addon | File>, file: any){
-    if (file.fileName){
-      const mappedEntity = fileMapper(file as IncomingFile)
-      const isExisting = await this.getById(File, file.id)
-      await this.orm.connect()
-      if (!isExisting){
-        this.entityManager.persist(mappedEntity)
-      } else if (isExisting){
-        console.log("Entity already exists with this id. Replace with new...")
-        this.entityManager.assign(mappedEntity, isExisting)
+  public async writeEntity(type: any, file: any) {
+    if (file.fileName) {
+      const mappedEntity = fileMapper(file as IncomingFile);
+      const isExisting = await this.getById(type, file.id);
+      await this.orm.connect();
+      if (!isExisting) {
+        this.entityManager.persist(mappedEntity);
+      } else if (isExisting) {
+        console.log("Entity already exists with this id. Replace with new...");
+        this.entityManager.assign(mappedEntity, isExisting);
       }
-    } else if (file.webSiteURL){
-      const mappedEntity = addonMapper(file as IncomingAddon)
-      const isExisting = await this.getById(Addon, file.addonID)
-      await this.orm.connect()
-      if (!isExisting){
-        this.entityManager.persist(mappedEntity)
-      } else if (isExisting){
-        console.log("Entity already exists with this id. Replace with new...")
-        this.entityManager.assign(mappedEntity, isExisting)
+    } else if (file.webSiteURL) {
+      const mappedEntity = addonMapper(file as IncomingAddon);
+      const isExisting = await this.getById(type, file.addonID);
+      await this.orm.connect();
+      if (!isExisting) {
+        this.entityManager.persist(mappedEntity);
+      } else if (isExisting) {
+        console.log("Entity already exists with this id. Replace with new...");
+        this.entityManager.assign(mappedEntity, isExisting);
       }
     }
-    await this.entityManager.flush()
+    await this.entityManager.flush();
     await this.orm.close();
   }
-
-  public getEntityManager() {
-    return this.entityManager;
-  }
-
 }
 
-export const dbORM = new AddonDB()
+export const dbORM = new AddonDB();
